+++ Assets/Scripts/Gameplay/Unit/CombatStrategies/IAttackStrategy.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public interface IAttackStrategy
 {
-    void ExecuteAttack(IDamageable target, float damage, Transform attackerTransform);
+    void ExecuteAttack(IDamageable target, int damage, Transform attackerTransform);
     bool CanAttack(float lastAttackTime, float attackCooldown);
 }
